+++ app/Http/Controllers/TravisController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\AlgoliaKeys;
+use App\Http\Middleware\IsTravisRunning;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 
@@ -15,6 +16,8 @@ class TravisController extends Controller
 
     public function __construct(AlgoliaKeys $algoliaKeys)
     {
+        $this->middleware(IsTravisRunning::class);
+
         $this->algoliaKeys = $algoliaKeys;
     }
